Entity 파일 정리

+++ src/main/java/com/education/onlinecampus/service/business/manager/CourseService.java
@@ -8,7 +8,7 @@ public interface CourseService {
 
     void CourseDelete(Course course);
 
-    Course CourseFind(Integer CourseSeq);
+    Course CourseFind(Long CourseSeq);
 
     void CourseChapterSave(CourseChapter courseChapter);
